+++ CHANGES.rst
@@ -5,6 +5,7 @@ nmea-kindle-panel changelog
 
 in progress
 ===========
+- UDP receiver based on ``asyncio-dgram`` and ``pynmea2``
 
 
 2022-07-xx 0.0.0